pyproject.toml
[project]
name = "simple-rag-project"
version = "0.1.0"
description = ""
authors = [
    { name = "Mykyta Lozhevych", email = "m.lozhevych@setuniversity.edu.ua" }
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "qdrant-client (>=1.14.3,<2.0.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flasgger (>=0.9.7,<0.10.0)",
    "gunicorn (>=21.2.0,<22.0.0)",
    "huggingface-hub (>=0.33.0,<0.34.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "openai (>=1.90.0,<2.0.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "python-json-logger (>=2.0.7,<3.0.0)",
    "opentelemetry-api (>=1.25.0,<2.0.0)",
    "opentelemetry-sdk (>=1.25.0,<2.0.0)",
    "opentelemetry-instrumentation-flask (>=0.46b0,<1.0.0)",
    "opentelemetry-exporter-otlp-proto-http (>=1.25.0,<2.0.0)",
    "prometheus-flask-exporter (>=0.23.0,<1.0.0)"
]

[tool.poetry]
packages = [{ include = "simple_rag_project", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create_embeddings = "simple_rag_project.create_embeddings:main"
